REMASTER COMEÇOU VAMO QUE VAMO
Decidi dar um UP no projeto pra matar a saudade desse projetinho e organizar ele tambem
Melhorias:
-Arquivo separado para funções costméticas
-Remake das funções mal-feitas
-Comentário revisados
- Bugs pequenos corrigidos

E se possivel quero deixar o programa mais bonito no geral
O foco é fazer ele rodar no Windows 10 e usando o DevC++ Embarcadeiro, mas se possivel irei preparar comentário para retrocompatibilidade
