+++ scripts/functions/desenhos.h
@@ -1,0 +1,765 @@
+/*Arquivo fonte para funções de desenhos do programa como a borda e os icones de:
+	-Disquete
+	-Logo
+	-Lupa
+	-CPU							11/04/2021 - Por quê refatorar esse programa antigo e mal feito?
+	-GPU							R: Saudades de mexer com macaco++ e começar algo do zero da trabalho
+	-RAM
+	-Mobo
+	-HDD
+*/
+
+// Definição de cores 
+#define cor_fundo 3
+#define cor_borda 1
+#define cor_texto 15
+#define cor_destaque 0
+#define RED 12
+
+void gotoxy(int x, int y);
+void textcolor(int newcolor);
+void textbackground(int newcolor);
+
+//Desenhos vão ficar todos aqui
+
+void borda() 
+{
+	int cf = cor_fundo, cb = cor_borda, lc = 160; // (cor de fundo, cor da borda, limite de coluna)
+	
+	textbackground(cf); // Define cor do fundo
+	system("cls"); 		//limpa sempre a tela quando é chamada
+	
+	for (int i = 9; i <= (lc - 8); i++) // Parte de cima da borda
+	{
+		textcolor(cb);
+		gotoxy(i,2);printf("%c",219);
+	}
+	for (int i = 9; i <= (lc - 8); i++) // Parte de baixo da borda 
+	{
+		textcolor(cb);
+		gotoxy(i,33);printf("%c",219);
+	}
+	for (int i=2; i <= 33; i++) // Borda esquerda 
+	{
+		textcolor(cb);
+		gotoxy(10,i);printf("%c",219);
+		gotoxy(9,i);printf("%c",219);
+	}
+	for (int i=2; i <= 33; i++) // Borda direita
+	{
+		textcolor(cb);
+		gotoxy(152,i);printf("%c",219);
+		gotoxy(151,i);printf("%c",219);
+	}
+	
+}
+
+void disquete(int ic, int il)
+{
+	
+	// Linha 1 => Borda de cima
+	textcolor(0);
+	gotoxy(ic, il); 
+	for(int i = 0; i < 20; i++) putchar(219);
+	
+	// Linha 2  - tracinho
+	gotoxy(ic, (il+1));
+	printf("%c", 219);
+	
+	for(int i = 0; i < 18; i++) 
+	{
+		if(i == 2 or i == 15) textcolor(8);
+		if(i > 2  and i < 15) 
+		{
+			textcolor(7);
+			if(i > 4 && i < 8) textcolor(0);
+		}
+		else textcolor(9);
+		putchar(219);
+	}
+	
+	textcolor(0);
+	printf("%c", 219);
+	
+	// Linha 3  - tracinho
+	gotoxy(ic, (il+2));
+	printf("%c", 219);
+	
+	for(int i = 0; i < 18; i++) 
+	{
+		if(i == 2 or i == 15) textcolor(8);
+		if(i > 2  and i < 15) 
+		{
+			textcolor(7);
+			if(i > 4 && i < 8) textcolor(0);
+		}
+		else textcolor(9);
+		putchar(219);
+	}
+	
+	textcolor(0);
+	printf("%c", 219);
+	
+	// Linha 4
+	
+	gotoxy(ic, (il+3));
+	
+	printf("%c", 219);
+	textcolor(9);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	
+	// Linha 5
+	
+	gotoxy(ic, (il+4));
+	printf("%c", 219);
+	textcolor(9);
+
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	
+	//Linha 6 - rodelinha com tracin
+	
+	gotoxy(ic, (il+5));
+	
+	printf("%c", 219);
+	for(int i = 0; i < 18; i++) 
+	{	
+		if(i>6 && i <11) 
+		{
+		textcolor(7);
+		if(i == 9) textcolor(8);
+		}
+		else textcolor(9);
+		putchar(219);
+		
+	}
+	textcolor(0);
+	printf("%c", 219);
+	
+	//Linha 7 - rodelinha 
+	
+	gotoxy(ic, (il+6));
+	
+	printf("%c", 219);
+	for(int i = 0; i < 18; i++) 
+	{	
+		if(i>6 && i <11) textcolor(7);
+		else textcolor(9);
+		putchar(219);
+		
+	}
+	textcolor(0);
+	printf("%c", 219);
+	
+	// Linha 8
+	
+	gotoxy(ic, (il+7));
+	
+	printf("%c", 219);
+	textcolor(9);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	
+	// Linha 9
+	
+	gotoxy(ic, (il+8));
+	
+	printf("%c", 219);
+	textcolor(9);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	
+	// Linha 10 => Borda de baixo
+	textcolor(0);
+	gotoxy(ic, (il+9)); 
+	for(int i = 0; i < 20; i++) putchar(219);
+	
+}
+
+void logo(int ic, int il)
+{
+	textcolor(0);
+	
+	gotoxy(ic - 2, il + 6); for(int i = 0; i <= 58; i++) printf("_");
+		
+	// Linha 1 do texto
+	gotoxy(ic+1,il);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);				gotoxy(ic+7,il);	printf("%c",187);
+	gotoxy(ic+10,il);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);				gotoxy(ic+16,il);	printf("%c",187);	
+	gotoxy(ic+18,il);	printf("%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219);	gotoxy(ic+25,il);	printf("%c",187);
+	gotoxy(ic+27,il);	printf("%c%c%c%c%c%c%c",219,219,219,219,219,219,219);		gotoxy(ic+34,il);	printf("%c",187);
+	gotoxy(ic+38,il);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);				gotoxy(ic+44,il);	printf("%c",187);
+	gotoxy(ic+46,il);	printf("%c%c",219,219);										gotoxy(ic+48,il);	printf("%c",187);
+	gotoxy(ic+51,il);	printf("%c%c",219,219);										gotoxy(ic+53,il);	printf("%c",187);
+	
+	//linha 2
+	gotoxy(ic,il+1);	printf("%c%c",219,219);		gotoxy(ic+2,il+1);	printf("%c%c%c%c%c%c",201,205,205,205,205,188);
+	gotoxy(ic+9,il+1);	printf("%c%c",219,219);		gotoxy(ic+11,il+1);	printf("%c%c%c%c",201,205,205,205);
+	gotoxy(ic+15,il+1);	printf("%c%c",219,219);		gotoxy(ic+17,il+1);	printf("%c",187);
+	gotoxy(ic+21,il+1);	printf("%c%c",219,219);		gotoxy(ic+18,il+1);	printf("%c%c%c",200,205,205);
+	gotoxy(ic+27,il+1);	printf("%c%c",219,219);		gotoxy(ic+23,il+1);	printf("%c%c%c",201,205,188);
+	gotoxy(ic+37,il+1);	printf("%c%c",219,219);		gotoxy(ic+29,il+1);	printf("%c%c%c%c%c%c",201,205,205,205,205,188);
+	gotoxy(ic+46,il+1);	printf("%c%c",219,219);		gotoxy(ic+39,il+1);	printf("%c%c%c%c%c%c",201,205,205,205,205,188);
+	gotoxy(ic+51,il+1);	printf("%c%c",219,219);		gotoxy(ic+48,il+1);	printf("%c",186);		gotoxy(ic+53,il+1);	printf("%c",186);
+		
+	//linha 3
+	gotoxy(ic,il+2);	printf("%c%c",219,219);									gotoxy(ic+2,il+2);	printf("%c",186);
+	gotoxy(ic+5,il+2);	printf("%c%c%c",219,219,219);							gotoxy(ic+8,il+2);	printf("%c",187);
+	gotoxy(ic+9,il+2);	printf("%c%c",219,219);									gotoxy(ic+11,il+2);	printf("%c",186);
+	gotoxy(ic+15,il+2);	printf("%c%c",219,219);									gotoxy(ic+17,il+2);	printf("%c",186);
+	gotoxy(ic+21,il+2);	printf("%c%c",219,219);									gotoxy(ic+23,il+2);	printf("%c",186);
+	gotoxy(ic+27,il+2);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);			gotoxy(ic+33,il+2);	printf("%c",187);
+	gotoxy(ic+37,il+2);	printf("%c%c",219,219);									gotoxy(ic+39,il+2);	printf("%c",186);
+	gotoxy(ic+46,il+2);	printf("%c%c%c%c%c%c%c",219,219,219,219,219,219,219);	gotoxy(ic+53,il+2);	printf("%c",186);
+	
+	//linha 4
+	gotoxy(ic,il+3);	printf("%c%c",219,219);		gotoxy(ic+2,il+3);	printf("%c",186);
+	gotoxy(ic+6,il+3);	printf("%c%c",219,219);		gotoxy(ic+8,il+3);	printf("%c",186);
+	gotoxy(ic+9,il+3);	printf("%c%c",219,219);		gotoxy(ic+11,il+3);	printf("%c",186);
+	gotoxy(ic+15,il+3);	printf("%c%c",219,219);		gotoxy(ic+17,il+3);	printf("%c",186);
+	gotoxy(ic+21,il+3);	printf("%c%c",219,219);		gotoxy(ic+23,il+3);	printf("%c",186);
+	gotoxy(ic+27,il+3);	printf("%c%c",219,219);		gotoxy(ic+29,il+3);	printf("%c%c%c%c%c",201,205,205,205,188);
+	gotoxy(ic+37,il+3);	printf("%c%c",219,219);		gotoxy(ic+39,il+3);	printf("%c",186);
+	gotoxy(ic+46,il+3);	printf("%c%c",219,219);		gotoxy(ic+48,il+3);	printf("%c%c%c",201,205,205);
+	gotoxy(ic+51,il+3);	printf("%c%c",219,219);		gotoxy(ic+53,il+3);	printf("%c",186);
+		
+	//linha 5
+	gotoxy(ic+1,il+4);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);			gotoxy(ic,il+4);	printf("%c",200);
+	gotoxy(ic+10,il+4);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);			gotoxy(ic+7,il+4);	printf("%c%c%c",201,188,200);
+	gotoxy(ic+21,il+4);	printf("%c%c",219,219);									gotoxy(ic+16,il+4);	printf("%c%c",201,188);
+	gotoxy(ic+27,il+4);	printf("%c%c%c%c%c%c%c",219,219,219,219,219,219,219);	gotoxy(ic+23,il+4);	printf("%c",186);
+	gotoxy(ic+38,il+4);	printf("%c%c%c%c%c%c",219,219,219,219,219,219);			gotoxy(ic+34,il+4);	printf("%c",187);
+	gotoxy(ic+46,il+4);	printf("%c%c",219,219);									gotoxy(ic+37,il+4);	printf("%c",200);
+	gotoxy(ic+51,il+4);	printf("%c%c",219,219);									gotoxy(ic+44,il+4);	printf("%c",187);
+	gotoxy(ic+48,il+4);	printf("%c",186);										gotoxy(ic+53,il+4);	printf("%c",186);
+	
+	//linha 6
+		
+	gotoxy(ic+1,il+5);	printf("%c%c%c%c%c%c%c",200,205,205,205,205,205,188);
+	gotoxy(ic+10,il+5);	printf("%c%c%c%c%c%c%c",200,205,205,205,205,205,188);
+	gotoxy(ic+21,il+5);	printf("%c%c%c",200,205,188);
+	gotoxy(ic+27,il+5);	printf("%c%c%c%c%c%c%c%c",200,205,205,205,205,205,205,188);
+	gotoxy(ic+38,il+5);	printf("%c%c%c%c%c%c%c",200,205,205,205,205,205,188);
+	gotoxy(ic+46,il+5);	printf("%c%c%c",200,205,188);
+	gotoxy(ic+51,il+5);	printf("%c%c%c",200,205,188);
+}
+
+void lupa(int ic, int il)
+{	
+	// Linha 1 - Borda Preta
+	textcolor(0);
+	// Linha 1 - Borda Preta
+	gotoxy(ic+4, il); 		printf("%c%c%c%c%c%c%c%c%c%c%c",220 ,220, 219, 219, 219, 219, 219, 219, 219, 220, 220);	
+	// Linha 2  
+	gotoxy(ic+2, il+1); 	printf("%c%c%c%c", 220, 219, 219, 223);
+	gotoxy(ic+13, il+1); 	printf("%c%c%c%c", 223, 219, 219, 220);
+	//linha 3
+	gotoxy(ic+1, il+2); 	printf("%c%c", 219, 223);
+	gotoxy(ic+16, il+2); 	printf("%c%c", 223, 219);	
+	//linha 4
+	gotoxy(ic, il+3);  		printf("%c%c", 219, 223);
+	gotoxy(ic+17, il+3); 	printf("%c%c", 223, 219);	
+	//linha 5
+	gotoxy(ic, il+4);	 	printf("%c", 219);
+	gotoxy(ic+18, il+4);	printf("%c", 219);	
+	//linha 6
+	gotoxy(ic, il+5);  		printf("%c%c", 219, 220);
+	gotoxy(ic+17, il+5);	printf("%c%c", 220, 219);	
+	//linha 7
+	gotoxy(ic+1, il+6);		printf("%c%c", 219, 220);
+	gotoxy(ic+16, il+6);	printf("%c%c", 220, 219);	
+	// Linha 8  
+	gotoxy(ic+2, il+7); 	printf("%c%c%c%c", 223, 219, 219, 220);
+	gotoxy(ic+13, il+7); 	printf("%c%c%c%c", 220, 219, 219, 223);
+	// Linha 9 
+	gotoxy(ic+4, il+8); 	printf("%c%c%c%c%c%c%c%c%c%c%c",223 ,223, 219, 219, 219, 219, 219, 219, 219, 223, 223);		
+	
+	//brilhinho
+	textcolor(15);
+	gotoxy(ic+3, il+4); printf("%c", 219);
+	gotoxy(ic+4, il+5); printf("%c%c%c", 219, 219, 220);
+	gotoxy(ic+7, il+6); printf("%c%c", 219, 220);
+	
+	//cabo
+	textcolor(15);
+	gotoxy(ic+18, il+6); printf("%c%c", 219,220);
+	gotoxy(ic+16, il+7); printf("%c%c%c%c%c%c", 219, 219, 219, 219, 219, 220);
+	gotoxy(ic+15, il+8); printf("%c%c%c%c%c%c%c%c%c", 223, 223, 219, 219, 219, 219, 219, 219,220);
+	gotoxy(ic+18, il+9); printf("%c%c%c%c%c%c%c%c", 223, 219, 219, 219, 219, 219, 219, 220);			//cabo
+	gotoxy(ic+20, il+10); printf("%c%c%c%c%c%c%c%c", 223, 219, 219, 219, 219, 219, 219, 220);
+	gotoxy(ic+22, il+11);printf("%c%c%c%c%c%c%c", 223, 219, 219, 219, 219, 219, 219);
+	gotoxy(ic+24, il+12);printf("%c%c%c%c%c", 223, 219, 219, 219, 219,223);
+	
+	// Acabou :D
+}
+
+void cpu(int ic, int il)
+{	
+	// Linha 1 => Borda de cima
+	textcolor(0);
+	gotoxy(ic, il); 
+	for(int i = 0; i < 20; i++) putchar(219);
+	
+	// Linha 2  - tracinho
+	gotoxy(ic, (il+1));
+
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	// Linha 3  - tracinho
+	gotoxy(ic, (il+2));
+	
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	// Linha 4
+	
+	gotoxy(ic, (il+3));
+	
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	// Linha 5
+	
+	gotoxy(ic, (il+4));
+	
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	//Linha 6 - rodelinha com tracin
+	
+	gotoxy(ic, (il+5));
+	
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	//Linha 7 - rodelinha 
+	
+	gotoxy(ic, (il+6));
+	
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	// Linha 8
+	
+	gotoxy(ic, (il+7));
+	
+	printf("%c", 219);
+	textcolor(7);
+	for(int i = 0; i < 18; i++) putchar(219);
+	textcolor(0);
+	printf("%c", 219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+		
+	// Linha 9 
+	textcolor(0);
+	gotoxy(ic, (il+8)); 
+	for(int i = 0; i < 20; i++) putchar(219);
+	textcolor(6);
+	printf("%c", 219);
+	textcolor(0);
+	
+	// Linha 10 => Borda de baixo
+	textcolor(6);
+	gotoxy((ic+1), (il+9)); 
+	for(int i = 0; i < 20; i++) putchar(219);
+	textcolor(0);
+	
+	//escrever cpu
+	gotoxy((ic+3), (il+2)); printf("%c%c%c%c",219,219,219,219);
+	gotoxy((ic+3), (il+3)); printf("%c",219);
+	gotoxy((ic+3), (il+4)); printf("%c",219);      //etra C
+	gotoxy((ic+3), (il+5)); printf("%c",219);
+	gotoxy((ic+3), (il+6)); printf("%c%c%c%c",219,219,219,219);
+	
+	gotoxy((ic+8), (il+2)); printf("%c%c%c%c",219,219,219,219);
+	gotoxy((ic+8), (il+3)); printf("%c",219);		//Letra P
+	gotoxy((ic+11), (il+3)); printf("%c",219);
+	gotoxy((ic+8), (il+4)); printf("%c%c%c%c",219,219,219,219);
+	gotoxy((ic+8), (il+5)); printf("%c",219);
+	gotoxy((ic+8), (il+6)); printf("%c",219);
+	
+	gotoxy((ic+13), (il+2)); printf("%c",219);
+	gotoxy((ic+16), (il+2)); printf("%c",219);
+	gotoxy((ic+13), (il+3)); printf("%c",219);
+	gotoxy((ic+16), (il+3)); printf("%c",219);		//Letra U
+	gotoxy((ic+13), (il+4)); printf("%c",219);
+	gotoxy((ic+16), (il+4)); printf("%c",219);
+	gotoxy((ic+13), (il+5)); printf("%c",219);
+	gotoxy((ic+16), (il+5)); printf("%c",219);
+	gotoxy((ic+13), (il+6)); printf("%c%c%c%c",219,219,219,219);
+	
+	gotoxy((ic), (il+9)); //só pro end do programa não quebrar as coisas
+
+}
+
+void gpu(int ic, int il)
+{	
+	// Linha 1 => Borda de cima
+	gotoxy(ic, il);
+	textcolor(8);
+	printf("%c%c%c", 219, 219, 219);
+	textcolor(7);	
+	for(int i = 0; i < 35; i++) putchar(219);
+	
+	// Linha 2  - começo do cooler
+	gotoxy(ic, (il+1));
+	textcolor(8);
+	printf("  %c", 219);
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(7);
+		if( i>14 && i<25)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	// Linha 3  
+	gotoxy(ic, (il+2));
+	
+	textcolor(8);
+	printf("  %c", 219);
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(2);
+		if( i>12 && i<27)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	// Linha 4
+	
+	gotoxy(ic, (il+3));
+	
+	textcolor(8);
+	printf("  %c", 219);
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(10);
+		if( i>11 && i<28)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	// Linha 5	- meio do cooler
+	gotoxy(ic, (il+4));
+	
+	textcolor(8);
+	printf("  %c", 219);	
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(10);
+		if( i>10 && i<29)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	//Linha 6 - 
+	
+	gotoxy(ic, (il+5));
+	
+	textcolor(8);
+	printf("  %c", 219);
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(10);
+		if( i>11 && i<28)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	//Linha 7
+	
+	gotoxy(ic, (il+6));
+	
+	textcolor(8);
+	printf("  %c", 219);
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(2);
+		if( i>12 && i<27)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	// Linha 8 - final do cooler 
+	
+	gotoxy(ic, (il+7));
+	
+	textcolor(8);
+	printf("  %c", 219);
+	for(int i = 0; i < 35; i++)
+	{
+		textcolor(7);
+		if( i>14 && i<25)
+			textcolor(8);
+		putchar(219);
+	}
+	
+	// Linha 9
+	
+	gotoxy(ic, (il+8));
+	
+	textcolor(8);
+	printf("  %c", 219);
+	textcolor(7);	
+	for(int i = 0; i < 35; i++) putchar(219);
+	
+	// Linha 10 => Borda de baixo
+	gotoxy(ic, (il+9)); 
+	
+	textcolor(8);
+	printf("  %c", 219);
+	textcolor(14);	
+	for(int i = 0; i < 35; i++) 
+		{
+			if(i==4 || i == 18)
+			{
+				textcolor(0);
+				putchar(219);
+				textcolor(14);
+			}
+			if(i>3 && i<18)
+			putchar(219);
+			else
+			printf(" ");
+		}
+	textcolor(0);
+	// Acabou :D
+}
+
+void ram(int ic, int il)
+{	
+	// Linha 1 => Borda de cima
+	gotoxy(ic, il); 
+	
+	textcolor(0); printf("%c", 219);
+	
+	textcolor(2);
+	for(int i = 0; i < 41; i++) putchar(219);
+	
+	textcolor(0); printf("%c", 219);
+	
+	// Linha 2  - tracinho
+	gotoxy(ic, (il+1));
+	
+	textcolor(0); printf(" %c", 219);
+		
+	for(int i = 0; i < 33; i++) 
+	{
+		textcolor(2);
+		if((i+1)%5==0) 
+		{
+			textcolor(7);
+			putchar(219);
+		}	
+		putchar(219);
+	}
+	
+	textcolor(0); printf("%c ", 219);
+	
+	// Linha 3  - tracinho
+	gotoxy(ic, (il+2));
+	
+	textcolor(0); printf("%c", 219);
+	
+	textcolor(2);
+	putchar(219);
+	for(int i = 0; i < 33; i++) 
+	{
+		textcolor(2);
+		if((i+1)%5==0) 
+		{
+			textcolor(7);
+			putchar(219);
+		}	
+		putchar(219);
+	}
+	putchar(219);
+	textcolor(0); printf("%c", 219);
+	
+	// Linha 4
+	
+	gotoxy(ic, (il+3));
+	
+	textcolor(0); printf(" %c", 219);
+	
+	textcolor(2);
+	for(int i = 0; i < 39; i++) putchar(219);
+	
+	textcolor(0); printf("%c ", 219);
+	
+	// Linha 5
+	
+	gotoxy(ic, (il+4));
+	textcolor(0); printf("%c", 219);
+	
+	for(int i = 0; i < 41; i++) 
+	{
+		textcolor(2);
+		if((i+1)%2==0)
+		textcolor(14);
+		putchar(219);
+	}
+	
+	textcolor(0); printf("%c", 219);
+	
+}
+
+void mobo(int ic, int il)
+{
+	// Construir o PCB bruto
+	gotoxy(ic, il); 
+	textcolor(0);	
+	for(int i = 0; i < 20; i++) 
+	{	
+		for(int j = 0; j < 10; j++)
+		{
+		gotoxy((ic+i), (il+j));
+		putchar(219);
+		}
+	}
+	
+	//Construir os detalhes
+	textcolor(7);
+	gotoxy(ic, (il+1)); printf("%c%c",219,219);
+	gotoxy(ic, (il+2)); printf("%c%c",219,219);
+	gotoxy(ic, (il+3)); printf("%c%c",219,219);		//IO traseiro
+	gotoxy(ic, (il+4)); printf("%c%c",219,219);
+	gotoxy(ic, (il+5)); printf("%c%c",219,219);
+	
+	textcolor(7);
+	gotoxy((ic+6), (il+2)); printf("%c",219);
+	gotoxy((ic+6), (il+3)); printf("%c",219);
+	gotoxy((ic+9), (il+2)); printf("%c",219);		//socket do cpu
+	gotoxy((ic+9), (il+3)); printf("%c",219);
+	textcolor(14);
+	gotoxy((ic+7), (il+2)); printf("%c%c",219,219);
+	gotoxy((ic+7), (il+3)); printf("%c%c",219,219);
+	
+	textcolor(2);
+	gotoxy((ic+13), (il+1)); printf("%c",219);
+	gotoxy((ic+13), (il+2)); printf("%c",219);
+	gotoxy((ic+13), (il+3)); printf("%c",219);    //slots de ram e etc
+	gotoxy((ic+13), (il+4)); printf("%c",219);
+	gotoxy((ic+15), (il+1)); printf("%c",219);
+	gotoxy((ic+15), (il+2)); printf("%c",219);
+	gotoxy((ic+15), (il+3)); printf("%c",219);
+	gotoxy((ic+15), (il+4)); printf("%c",219);	
+	gotoxy((ic+18), (il+1)); printf("%c%c",219,219);
+	
+	textcolor(8);
+	gotoxy((ic+18), (il+3)); printf("%c%c",219,219);    //conectores 24 pinos e etc
+	gotoxy((ic+18), (il+4)); printf("%c%c",219,219);
+	
+	gotoxy((ic+4), (il+6)); printf("%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219);  //slots pci-e
+	gotoxy((ic+4), (il+8)); printf("%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219);
+	
+	textcolor(RED);
+	gotoxy((ic+15), (il+6)); printf("%c%c",219,219); //dissipador do chipset
+	gotoxy((ic+15), (il+7)); printf("%c%c",219,219);
+	
+	textcolor(15);
+	gotoxy((ic+14), (il+9)); printf("%c%c%c%c",219,219,219,219);
+	textcolor(0);
+	
+	gotoxy((ic+10), (il+10));	// tirar o exit do programa de perto
+}
+
+void HDD(int ic, int il)
+{
+	textcolor(0);
+	for(int i = 0; i < 21; i++) 
+	{	
+		for(int j = 0; j < 12; j++)
+		{
+		gotoxy((ic+i), (il+j));
+		putchar(219);
+		}
+	}
+	textbackground(0);
+	textcolor(7);
+	gotoxy(ic+6,il+1);		printf("%c%c%c%c%c%c%c%c%c",220,220,219,219,219,219,219,220,220);
+	gotoxy(ic+4,il+2);		printf("%c%c%c%c%c%c%c%c%c%c%c%c%c",220,219,219,219,219,219,219,219,219,219,219,219,220);
+	gotoxy(ic+4,il+3);		printf("%c%c%c%c%c%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219,219,219,219,219,219);		//disco magnético
+	gotoxy(ic+4,il+4);		printf("%c%c%c%c%c%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219,219,219,219,219,219);		//e parafusos
+	gotoxy(ic+5,il+5);		printf("%c%c%c%c%c%c%c%c%c%c%c",223,219,219,219,219,219,219,219,219,219,223);
+	gotoxy(ic+8,il+6);		printf("%c%c%c%c%c",223,223,223,223,223);
+	gotoxy(ic+1,il);		printf("%c",220);	gotoxy(ic+1,il+11);		printf("%c",223);
+	gotoxy(ic+1,il+5);		printf("%c",223);	gotoxy(ic+19,il+5);		printf("%c",223);
+	gotoxy(ic+19,il);		printf("%c",220);	gotoxy(ic+19,il+11);	printf("%c",223);
+		
+	textcolor(2);
+	gotoxy(ic+10,il+7);		printf("%c%c%c%c%c%c%c%c",220,220,220,220,220,220,220,220);
+	gotoxy(ic+10,il+8);		printf("%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219); // pcb
+	gotoxy(ic+12,il+9);		printf("%c%c%c%c%c%c",219,219,219,219,219,219);
+	gotoxy(ic+13,il+10);	printf("%c%c%c%c%c",219,219,219,219,219);
+	textcolor(8);
+	gotoxy(ic+14,il+8);		printf("%c%c%c",219,219,219);
+	gotoxy(ic+15,il+9);		printf("%c%c",219,219);
+	
+	textcolor(7);
+	gotoxy(ic+3,il+6);		printf("%c",220);
+	gotoxy(ic+3,il+7);		printf("%c%c%c",219,219,220);										//baia para a agulha
+	gotoxy(ic+3,il+8);		printf("%c%c%c%c%c",219,219,219,219,220);
+	gotoxy(ic+3,il+9);		printf("%c%c%c%c%c%c%c%c",219,219,219,219,219,219,220,220);
+	gotoxy(ic+3,il+10);		printf("%c%c%c%c%c%c%c%c%c%c%c%c",219,219,219,219,219,219,219,219,219,219,220,220);
+	
+	textcolor(8);
+	textbackground(7);
+	//gotoxy(ic+9,il+4);		printf("%c",220);
+	gotoxy(ic+9,il+5);		printf("%c",219);
+	gotoxy(ic+8,il+6);		printf("%c%c",219,219);	
+	gotoxy(ic+7,il+7);		printf("%c%c%c",219,219,219);			//agulha
+	gotoxy(ic+6,il+8);		printf("%c%c%c",219,219,219);	
+	gotoxy(ic+5,il+9);		printf("%c%c%c%c",219,219,219,219);
+	gotoxy(ic+6,il+10);		printf("%c%c",223,223);	
+	
+	textcolor(0);
+	gotoxy(ic+9,il+3);		printf("%c%c%c",220,219,220);		///motor para o disco
+	gotoxy(ic+10,il+4);		printf("%c",223);
+	
+	textbackground(cor_fundo);
+}
